refactor: Moved auth strategies into folder #111
