+++ apps/mull-api/src/app/auth/strategies/twitter.strategy.ts
@@ -1,9 +1,9 @@
 import { PassportStrategy } from '@nestjs/passport';
 import { IStrategyOption, Profile, Strategy } from 'passport-twitter';
-import { environment } from '../../environments/environment';
+import { environment } from '../../../environments/environment';
 import { Injectable } from '@nestjs/common';
-import { User } from '../entities';
-import { AuthService } from './auth.service';
+import { User } from '../../entities';
+import { AuthService } from './../auth.service';
 import { RegistrationMethod } from '@mull/types';
 
 @Injectable()
